model fine-tuning
Summary:
The first part of URU recipe trains the model on large Instagram hashtag dataset, the second step is to fine-tuned the pre-trained model on ImageNet to evaluate model performance.
This diff implements the model fine-tuning part.
We wrap the pre-trained model with FineTuningWrapper, which uses feature extractor to obtain embeddings from the trunk. The wrapper forwards the embeddings to a trainable FullyConnected head to train the classifier.

Reviewed By: sallysyw

Differential Revision: D32153843

fbshipit-source-id: c8f69be4fb696bfb71269ff478c5e2bfdecaaa5c