include module path to API usage log
Summary:
as suggested by kazhang, there could be multiple apps with the same class name but in different module. e.g. there are 2 versions of doc classification in torchrecipes.

This diff add module path to the event name in the usage logging to ensure we can distinguish the usage of train apps of above case.

Caveat:
As we prepend "torchrecipes" to the event name as a search pattern in query. There will be redundant "torchrecipes" in the event name for train apps inside of torchrecipes package. Otherwise, it would be difficult to capture train apps outside of torchrecipes package. See the test plan for examples

Reviewed By: tangbinh

Differential Revision: D32962688

fbshipit-source-id: e7b4b2e12c96f7019684f14cffa192631ffbcf26