Remove torch, torchvision, torchtext, torchdata from requirements.txt
Summary: As requested, we remove `torch` and some domain libraries from `requirement.txt`. It's recommended that these libraries are installed via `conda` instead of `pip`, which is consistent with the lists of requirements from `torchvision` and `torchtext`. In addition, we avoid having to specify CPU versions of these libraries, which might force users to reinstall them in order to run GPU jobs.

Reviewed By: kalyanvasudev

Differential Revision: D33540170

fbshipit-source-id: d02a99783f0cd0907bd5a7ce5d4b8eadb438db9f