Cosine with warmup LR scheduler
Summary:
This diff introduces a Cosine with warmup LR scheduler to TorchRecipes. The LR scheduler is essentially the same as `stl.lightning.optim.lr_schedulers.cosine_with_warmup`, but build on top of torch LR schedulers instead of home-cook implementation.
As a follow-up, we can replace the STL one with this implementation (the interface is slightly different, so we might have to update configs). cc tangbinh

* Why not directly use `torch.optim.lr_scheduler.SequentialLR`? which allows you to mix match any LR schedulers in config
  * It's very cumbersome to compose that in Hydra config, and we have to manually calculate the milestones;
https://www.internalfb.com/code/fbsource/[b49434434effaec39dfb5980e8d74609746e262d]/fbcode/torchrecipes/fb/projects/uru/conf/trunk_training_8_nodes.yaml?lines=30-43

Reviewed By: hudeven

Differential Revision: D33729078

fbshipit-source-id: db8a752504bbaf95ece1f7c913c14009f04b6816