Remove FineTuningClassyConf, HiveWritingClassyConf, and MixupTransformConf
Summary: We remove three dataclasses for callbacks, namely `FineTuningClassyConf`, `HiveWritingClassyConf`, and `MixupTransformConf`. The registration code for these dataclasses with `ConfigStore` has been dead for a while, so we remove it as well. The refactor is part of the V2 migration for `torchrecipes/fb/vision/image_classification`, which has been broken and needs to be fixed (cc hazirbas).

Reviewed By: hudeven

Differential Revision: D35740137

fbshipit-source-id: 3fecde9586e58fefea95ae353cdbdad7ab84ed3b