Randomize num ids per feature
Summary:
X-link: https://github.com/pytorch/torchrec/pull/1348


To be a good citizen, I changed the default behavior to sample from 0 to num ids. This necessitated fixing some unit tests which had assumed that the number of ids would be fixed. I can also NOT change the default, but it seems like a good idea for the random dataset to actually have random ids by default.

Reviewed By: colin2328, YLGH

Differential Revision: D48569957